feat: Add skeleton button (#19699)
* #18237 add skeleton button

* update snapshot

* refactoring demo code

* lint code

* update test snapshots

* add test

* update snapshot

* fix deps lint issue

* set avatar siz and update test snapshots

* fix lint issue

* remove button and just keep skeleton

* add active switch for button demo

* add size and shape radio for skeleton button demo

* add button tests

* add skeleton tests

* update doc

* update test snapshots

* omit avatar and button props

* refactoring skeleton and update test snapshots
