+++ components/skeleton/Avatar.tsx
@@ -1,47 +1,18 @@
 import * as React from 'react';
-import classNames from 'classnames';
+import SkeletonElement, { SkeletonElementProps } from './SkeletonElement';
 
-export interface SkeletonAvatarProps {
-  prefixCls?: string;
-  className?: string;
-  style?: object;
-  size?: 'large' | 'small' | 'default' | number;
+export interface AvatarProps extends Omit<SkeletonElementProps, 'shape'> {
   shape?: 'circle' | 'square';
 }
 
 // eslint-disable-next-line react/prefer-stateless-function
-class SkeletonAvatar extends React.Component<SkeletonAvatarProps, any> {
-  static defaultProps: Partial<SkeletonAvatarProps> = {
+class SkeletonAvatar extends React.Component<AvatarProps, any> {
+  static defaultProps: Partial<SkeletonElementProps> = {
     size: 'large',
   };
 
   render() {
-    const { prefixCls, className, style, size, shape } = this.props;
-
-    const sizeCls = classNames({
-      [`${prefixCls}-lg`]: size === 'large',
-      [`${prefixCls}-sm`]: size === 'small',
-    });
-
-    const shapeCls = classNames({
-      [`${prefixCls}-circle`]: shape === 'circle',
-      [`${prefixCls}-square`]: shape === 'square',
-    });
-
-    const sizeStyle: React.CSSProperties =
-      typeof size === 'number'
-        ? {
-            width: size,
-            height: size,
-            lineHeight: `${size}px`,
-          }
-        : {};
-    return (
-      <span
-        className={classNames(prefixCls, className, sizeCls, shapeCls)}
-        style={{ ...sizeStyle, ...style }}
-      />
-    );
+    return <SkeletonElement {...this.props} />;
   }
 }
