+++ components/skeleton/style/index.less
@@ -5,6 +5,7 @@
 @skeleton-avatar-prefix-cls: ~'@{skeleton-prefix-cls}-avatar';
 @skeleton-title-prefix-cls: ~'@{skeleton-prefix-cls}-title';
 @skeleton-paragraph-prefix-cls: ~'@{skeleton-prefix-cls}-paragraph';
+@skeleton-button-prefix-cls: ~'@{skeleton-prefix-cls}-button';
 
 @skeleton-to-color: shade(@skeleton-color, 5%);
 
@@ -23,14 +24,14 @@
       vertical-align: top;
       background: @skeleton-color;
 
-      .avatar-size(@avatar-size-base);
+      .skeleton-element-avatar-size(@avatar-size-base);
 
       &-lg {
-        .avatar-size(@avatar-size-lg);
+        .skeleton-element-avatar-size(@avatar-size-lg);
       }
 
       &-sm {
-        .avatar-size(@avatar-size-sm);
+        .skeleton-element-avatar-size(@avatar-size-sm);
       }
     }
   }
@@ -96,17 +97,61 @@
     .@{skeleton-avatar-prefix-cls} {
       .skeleton-color();
     }
+
+    .@{skeleton-button-prefix-cls} {
+      .skeleton-color();
+    }
+  }
+
+  // Skeleton element
+  &-element {
+    display: inline-block;
+    // Button
+    .@{skeleton-button-prefix-cls} {
+      display: inline-block;
+      vertical-align: top;
+      background: @skeleton-color;
+      border-radius: @border-radius-base;
+
+      .skeleton-element-button-size(@btn-height-base);
+
+      &-lg {
+        .skeleton-element-button-size(@btn-height-lg);
+      }
+
+      &-sm {
+        .skeleton-element-button-size(@btn-height-sm);
+      }
+    }
   }
 }
 
-.avatar-size(@size) {
+.skeleton-element-avatar-size(@size) {
   width: @size;
-  height: @size;
-  line-height: @size;
+  .skeleton-element-common-size(@size);
 
   &.@{skeleton-avatar-prefix-cls}-circle {
     border-radius: 50%;
   }
+}
+
+.skeleton-element-button-size(@size) {
+  width: @size * 2;
+  .skeleton-element-common-size(@size);
+
+  &.@{skeleton-button-prefix-cls}-circle {
+    width: @size;
+    border-radius: 50%;
+  }
+
+  &.@{skeleton-button-prefix-cls}-round {
+    border-radius: @size;
+  }
+}
+
+.skeleton-element-common-size(@size) {
+  height: @size;
+  line-height: @size;
 }
 
 .skeleton-color() {
